requirements.txt
Cython
fastapi>=0.100.0
uvicorn>=0.22.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
Faker>=18.3.1
numpy>=1.25.0
pandas>=2.0.0
xgboost
imbalanced-learn>=0.10.1
# Privacy-aware fraud detection dependencies
redis>=4.5.0
kafka-python>=2.0.2
networkx>=3.0
shap>=0.42.0
scikit-learn>=1.3.0
aioredis>=2.0.1
asyncio-mqtt>=0.11.0
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0
websockets>=11.0
cryptography>=41.0.0
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.6
